JAMES-2287 Cassandra ObjectStore implementation
